Add multiple linear regression model for housing prices
Problem Statement
A real estate company wishes to analyse the prices of properties based on various
factors such as area, number of rooms, bathrooms, bedrooms, etc. Create a multiple
linear regression model which is capable of predicting the sale price of houses based on
multiple factors and evaluate the accuracy of this model.